pyproject.toml
[project]
name = "ai-engineer"
version = "0.1.0"
description = "AI Engineer Learning Repository"
requires-python = ">=3.12,<3.13"
dependencies = [
    # OpenAI & Anthropic
    "openai>=1.12.0",
    "anthropic>=0.18.1",
    
    # LiteLLM & Google
    "litellm>=1.30.0",
    "google-generativeai>=0.3.2",
    
    # Hugging Face Ecosystem
    "transformers>=4.37.2",
    "huggingface-hub>=0.20.3",
    "sentence-transformers>=2.3.1",
    "accelerate>=0.26.1",
    
    # Vector Databases
    "chromadb>=0.4.22",
    "pinecone>=8.0.0",
    "weaviate-client>=3.26.2",
    
    # AI Frameworks
    "langchain>=1.0.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.19",
    "llama-index>=0.9.48",
    
    # Token Management
    "tiktoken>=0.5.2",
    
    # Data Validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
    
    # HTTP & API
    "requests>=2.31.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.3",
    
    # Data Processing
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "faiss-cpu>=1.7.0",
    
    # Jupyter
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.11",
    "ipywidgets>=8.1.1",
    "notebook>=7.0.7",
    
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "pytest-cov>=4.1.0",
    
    # Utilities
    "tenacity>=8.2.3",
    "tqdm>=4.66.1",
    "colorama>=0.4.6",
    "python-json-logger>=2.0.7",
    
    # File Formats
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
